requirements.txt
Django==1.11.13
geopy==1.16.0
requests==2.19.1
Pillow==5.2.0
